+++ src/particles/dimensions/theme.ts
@@ -1,3 +1,8 @@
+import React from 'react';
+
+import { useWindowSize } from "@kibalabs/core-react";
+
+import { useDimensions } from '../../theming';
 import { ThemeType } from '../../util';
 
 export interface IDimensionGuide extends ThemeType {
@@ -31,6 +36,7 @@ export interface IDimensionGuide extends ThemeType {
 }
 
 export enum ScreenSize {
+  Base = 'base',
   Small = 'small',
   Medium = 'medium',
   Large = 'large',
@@ -89,3 +95,31 @@ export const getPaddingSize = (size: PaddingSizeProp, theme: IDimensionGuide): s
   console.error(`Failed to find padding size: ${size} (${fieldName})`);
   return '0px';
 };
+
+export const ResponsiveScreenSizeContext = React.createContext<ScreenSize>(ScreenSize.Base);
+
+const getScreenSizeValue = (size: ScreenSize, theme: IDimensionGuide): number => {
+  if (size === ScreenSize.Base) {
+    return 0;
+  }
+  return Number(getScreenSize(size as string as ScreenSize, theme).replace('px', ''));
+};
+
+export function useResponsiveScreenSize(): ScreenSize {
+  const dimensions = useDimensions();
+  const windowSize = useWindowSize();
+  let screenSize = ScreenSize.Base;
+  if (windowSize.width > getScreenSizeValue(ScreenSize.Small, dimensions)) {
+    screenSize = ScreenSize.Small;
+  }
+  if (windowSize.width > getScreenSizeValue(ScreenSize.Medium, dimensions)) {
+    screenSize = ScreenSize.Medium;
+  }
+  if (windowSize.width > getScreenSizeValue(ScreenSize.Large, dimensions)) {
+    screenSize = ScreenSize.Large;
+  }
+  if (windowSize.width > getScreenSizeValue(ScreenSize.ExtraLarge, dimensions)) {
+    screenSize = ScreenSize.ExtraLarge;
+  }
+  return screenSize;
+}